feat(tests): refactor testing framework with BaseTestable and MakesAssertions
- Introduced BaseTestable class to encapsulate common testing logic for components.
- Implemented ChainableTest for fluent asynchronous testing API.
- Refactored MakesAssertions to be a higher-order function that extends BaseTestable.
- Updated ComponentTest to utilize BaseTestable and MakesAssertions for improved structure and maintainability.
- Removed redundant code and streamlined assertion methods for better clarity and performance.
